wip-suffix-overload-removal

+++ packages/replicad/src/register.ts
@@ -16,6 +16,7 @@ if (!(globalThis as any).FinalizationRegistry) {
 
 const deletetableRegistry = new (globalThis as any).FinalizationRegistry(
   (heldValue: Deletable) => {
+    if (!heldValue) return;
     try {
       heldValue.delete();
     } catch (e) {
@@ -47,7 +48,9 @@ export class WrappingObj<Type extends Deletable> {
       this._wrapped.delete();
     }
 
-    deletetableRegistry.register(this, newWrapped, newWrapped);
+    if (newWrapped) {
+      deletetableRegistry.register(this, newWrapped, newWrapped);
+    }
     this._wrapped = newWrapped;
   }
 
@@ -60,7 +63,9 @@ export class WrappingObj<Type extends Deletable> {
 
 export const GCWithScope = () => {
   function gcWithScope<Type extends Deletable>(value: Type): Type {
-    deletetableRegistry.register(gcWithScope, value);
+    if (value) {
+      deletetableRegistry.register(gcWithScope, value);
+    }
     return value;
   }
 
@@ -69,7 +74,9 @@ export const GCWithScope = () => {
 
 export const GCWithObject = (obj: any) => {
   function registerForGC<Type extends Deletable>(value: Type): Type {
-    deletetableRegistry.register(obj, value);
+    if (value) {
+      deletetableRegistry.register(obj, value);
+    }
     return value;
   }
